dev: move steps test in test_e2e.py
